+++ src/graphics.rs
@@ -11,13 +11,12 @@
 
 use crate::QuicksilverError;
 
+mod circle_points;
 mod color;
-mod image;
 mod mesh;
 mod vertex;
 
 pub use self::color::Color;
-pub use self::image::Image;
 pub use self::mesh::Mesh;
 pub use self::vertex::{Element, Vertex};
 
@@ -26,9 +25,10 @@ use golem::*;
 
 pub use golem::ColorFormat as PixelFormat;
 
+use std::cell::{Ref, RefCell};
 use std::iter;
-
-// TODO: image views
+use std::path::Path;
+use std::rc::Rc;
 
 /// The struct that handles sending draw calls to the GPU
 ///
@@ -53,6 +53,7 @@ pub struct Graphics {
     image_changes: Vec<(usize, Image)>,
     projection_changes: Vec<(usize, Transform)>,
     geom_mode_changes: Vec<(usize, GeometryMode)>,
+    clear_changes: Vec<(usize, Color)>,
     transform: Transform,
 }
 
@@ -107,14 +108,15 @@ impl Graphics {
             image_changes: Vec::new(),
             projection_changes: Vec::new(),
             geom_mode_changes: Vec::new(),
+            clear_changes: Vec::new(),
             transform: Transform::IDENTITY,
         })
     }
 
     /// Clear the screen to the given color
     pub fn clear(&mut self, color: Color) {
-        self.ctx.set_clear_color(color.r, color.g, color.b, color.a);
-        self.ctx.clear();
+        let head = self.index_data.len();
+        self.clear_changes.push((head, color));
     }
 
     /// Set the projection matrix, which is applied to all vertices on the GPU
@@ -298,8 +300,8 @@ impl Graphics {
         self.stroke_polygon(&Self::circle_points(center, radius)[..], color);
     }
 
-    fn circle_points(center: Vector, radius: f32) -> [Vector; CIRCLE_LEN] {
-        let mut points = CIRCLE_POINTS;
+    fn circle_points(center: Vector, radius: f32) -> [Vector; circle_points::CIRCLE_LEN] {
+        let mut points = circle_points::CIRCLE_POINTS;
         for point in points.iter_mut() {
             *point = Vector {
                 x: center.x + radius * point.x,
@@ -377,9 +379,16 @@ impl Graphics {
 
     /// Send the accumulated draw data to the GPU
     ///
-    /// This should almost never be necessary for a user to call directly, use
-    /// [`Graphics::present`] instead.
-    pub fn flush(&mut self) -> Result<(), QuicksilverError> {
+    /// Except when rendering to a [`Surface`], this should almost never be necessary for a user
+    /// to call directly. Use [`Graphics::present`] to draw to the window instead. When rendering
+    /// to a [`Surface`], remember to set the viewport via [`Graphics::set_viewport`]
+    pub fn flush(&mut self, surface: Option<&Surface>) -> Result<(), QuicksilverError> {
+        // Either bind a surface or draw directly to the default framebuffer, depending on the
+        // argument
+        match surface {
+            Some(surface) => surface.0.bind(),
+            None => golem::Surface::unbind(&self.ctx),
+        }
         const TEX_BIND_POINT: u32 = 1;
         let max_index = (self.vertex_data.len() / VERTEX_SIZE) as u32;
         for index in self.index_data.iter() {
@@ -399,10 +408,13 @@ impl Graphics {
         let mut element_mode = GeometryMode::Triangles;
         let change_list = join_change_lists(
             join_change_lists(
-                self.image_changes.drain(..),
-                self.projection_changes.drain(..),
+                join_change_lists(
+                    self.image_changes.drain(..),
+                    self.projection_changes.drain(..),
+                ),
+                self.geom_mode_changes.drain(..),
             ),
-            self.geom_mode_changes.drain(..),
+            self.clear_changes.drain(..),
         );
         for (index, changes) in change_list {
             // Before we change state, draw the old state
@@ -413,22 +425,28 @@ impl Graphics {
                 previous = index;
             }
             // Change the render state
-            if let Some(first) = changes.0 {
-                // If we're switching what image to use, do so now
-                if let Some(image) = first.0 {
-                    let bind_point = std::num::NonZeroU32::new(TEX_BIND_POINT).unwrap();
-                    image.raw().set_active(bind_point);
+            if let Some(changes) = changes.0 {
+                if let Some(changes) = changes.0 {
+                    // If we're switching what image to use, do so now
+                    if let Some(image) = changes.0 {
+                        let bind_point = std::num::NonZeroU32::new(TEX_BIND_POINT).unwrap();
+                        image.raw().set_active(bind_point);
+                    }
+                    // If we're switching what projection to use, do so now
+                    if let Some(projection) = changes.1 {
+                        let matrix = Self::transform_to_gl(projection);
+                        self.shader
+                            .set_uniform("projection", UniformValue::Matrix3(matrix))?;
+                    }
                 }
-                // If we're switching what projection to use, do so now
-                if let Some(projection) = first.1 {
-                    let matrix = Self::transform_to_gl(projection);
-                    self.shader
-                        .set_uniform("projection", UniformValue::Matrix3(matrix))?;
+                // If we're switching the element mode, do so now
+                if let Some(g_m) = changes.1 {
+                    element_mode = g_m;
                 }
             }
-            // If we're switching the element mode, do so now
-            if let Some(g_m) = changes.1 {
-                element_mode = g_m;
+            if let Some(color) = changes.1 {
+                self.ctx.set_clear_color(color.r, color.g, color.b, color.a);
+                self.ctx.clear();
             }
         }
         if previous != self.index_data.len() {
@@ -437,6 +455,7 @@ impl Graphics {
                     .draw_prepared(previous..self.index_data.len(), element_mode);
             }
         }
+        golem::Surface::unbind(&self.ctx);
         self.vertex_data.clear();
         self.index_data.clear();
 
@@ -453,23 +472,206 @@ impl Graphics {
 
     /// Send the draw data to the GPU and paint it to the Window
     pub fn present(&mut self, win: &blinds::Window) -> Result<(), QuicksilverError> {
-        self.flush()?;
+        self.flush(None)?;
         win.present();
 
         Ok(())
     }
 
-    pub(crate) fn create_image(
-        &self,
-        data: &[u8],
+    /// Select the area to draw to
+    ///
+    /// Generally, the best practice is to set this to (0, 0, window_width, window_height), but
+    /// when rendering to a subset of the screen it can be useful to change this. Additionally, you
+    /// probably want to set the viewport when rendering to a [`Surface`] in [`Graphics::flush`].
+    ///
+    /// The units given are physical units, not logical units. As such when using [`Window::width`]
+    /// and [`Window::height`], be sure to multiply by [`Window::scale_factor`].
+    ///
+    /// [`Window::width`]: crate::lifecycle::Window::width
+    /// [`Window::height`]: crate::lifecycle::Window::height
+    /// [`Window::scale_factor`]: crate::lifecycle::Window::scale_factor
+    pub fn set_viewport(&self, x: u32, y: u32, width: u32, height: u32) {
+        self.ctx.set_viewport(x, y, width, height);
+    }
+}
+
+/// A 2D image, stored on the GPU
+///
+/// See [`Graphics::draw_image`] to draw it
+#[derive(Clone)]
+pub struct Image(Rc<RefCell<Texture>>);
+
+impl Image {
+    fn new(texture: Texture) -> Image {
+        Image(Rc::new(RefCell::new(texture)))
+    }
+
+    /// Create an image with a given width and height
+    ///
+    /// Either source the data from an array of bytes, or create a blank image.
+    /// `format` determines how to interpet the bytes when creating the image
+    pub fn from_raw(
+        gfx: &Graphics,
+        data: Option<&[u8]>,
         width: u32,
         height: u32,
         format: PixelFormat,
-    ) -> Result<Texture, GolemError> {
-        let mut texture = Texture::new(&self.ctx)?;
-        texture.set_image(Some(data), width, height, format);
+    ) -> Result<Image, GolemError> {
+        let mut texture = Texture::new(&gfx.ctx)?;
+        texture.set_image(data, width, height, format);
 
-        Ok(texture)
+        Ok(Image::new(texture))
+    }
+
+    /// Create an image from an encoded image format
+    ///
+    /// JPEG and PNG are supported
+    pub fn from_encoded_bytes(gfx: &Graphics, raw: &[u8]) -> Result<Image, QuicksilverError> {
+        let img = image::load_from_memory(raw)?.to_rgba();
+        let width = img.width();
+        let height = img.height();
+        Ok(Image::from_raw(
+            gfx,
+            Some(img.into_raw().as_slice()),
+            width,
+            height,
+            PixelFormat::RGBA,
+        )?)
+    }
+
+    /// Load an image from a file at the given path
+    ///
+    /// JPEG and PNG file formats are supported
+    pub async fn load(gfx: &Graphics, path: impl AsRef<Path>) -> Result<Image, QuicksilverError> {
+        let file_contents = platter::load_file(path).await?;
+        Image::from_encoded_bytes(gfx, file_contents.as_slice())
+    }
+
+    /// Replace the backing data for the image, or create a blank image
+    pub fn set_data(&mut self, data: Option<&[u8]>, width: u32, height: u32, color: ColorFormat) {
+        self.0.borrow_mut().set_image(data, width, height, color);
+    }
+
+    /// Set the data for some region of this image, without clearing it
+    pub fn set_sub_data(
+        &self,
+        data: &[u8],
+        x: u32,
+        y: u32,
+        width: u32,
+        height: u32,
+        color: ColorFormat,
+    ) {
+        self.raw().set_subimage(data, x, y, width, height, color);
+    }
+
+    pub(crate) fn raw(&self) -> Ref<Texture> {
+        self.0.borrow()
+    }
+
+    pub(crate) fn ptr_eq(&self, other: &Image) -> bool {
+        Rc::ptr_eq(&self.0, &other.0)
+    }
+
+    /// Get the size of the image
+    pub fn size(&self) -> Vector {
+        Vector {
+            x: self.raw().width() as f32,
+            y: self.raw().height() as f32,
+        }
+    }
+
+    /// Determine how the texture should scale down
+    pub fn set_minification(&self, min: TextureFilter) {
+        self.raw().set_minification(min);
+    }
+
+    /// Determine how the texture should scale up
+    pub fn set_magnification(&self, max: TextureFilter) {
+        self.raw().set_magnification(max);
+    }
+
+    /// Determine how the texture is wrapped horizontally
+    pub fn set_wrap_h(&self, wrap: TextureWrap) {
+        self.raw().set_wrap_h(wrap);
+    }
+
+    /// Determine how the texture is wrapped vertically
+    pub fn set_wrap_v(&self, wrap: TextureWrap) {
+        self.raw().set_wrap_v(wrap);
+    }
+
+    fn into_raw(self) -> Result<Texture, Rc<RefCell<Texture>>> {
+        Ok(Rc::try_unwrap(self.0)?.into_inner())
+    }
+}
+
+/// A Surface is the core struct for rendering to textures, or getting data from them.
+///
+/// If you want to render to a texture, [`attach`] it and then pass the surface to
+/// [`Graphics::flush`].
+///
+/// If you want to get data from a texture, [`attach`] it and use [`Surface::screenshot`].
+///
+/// [`attach`]: Surface::attach
+pub struct Surface(golem::Surface);
+
+impl Surface {
+    /// Create a Surface with an attached Image
+    ///
+    /// The image must not have any other references to it, or this function will return an error.
+    pub fn new(gfx: &Graphics, attachment: Image) -> Result<Surface, QuicksilverError> {
+        let tex = attachment
+            .into_raw()
+            .map_err(|_| QuicksilverError::SurfaceImageError)?;
+        Ok(Surface(golem::Surface::new(&gfx.ctx, tex)?))
+    }
+
+    /// Use the attached image as the backing data for this Surface
+    ///
+    /// To either get the data for an image via [`Surface::screenshot`] or set it via
+    /// [`Graphics::flush`], an image needs to be attached to this Surface.
+    ///
+    /// The image must not have any other references to it, or this function will return an error.
+    ///
+    /// It's generally faster to create one [`Surface`] per [`Image`], and only attach and
+    /// [`detach`] when necessary.
+    ///
+    /// [`detach`]: Surface::detach
+    pub fn attach(&mut self, attachment: Image) -> Result<(), QuicksilverError> {
+        let tex = attachment
+            .into_raw()
+            .map_err(|_| QuicksilverError::SurfaceImageError)?;
+        self.0.put_texture(tex);
+
+        Ok(())
+    }
+
+    /// Take the Image out of this Surface
+    ///
+    /// To use the data that has been rendered to a Surface, its attachment has to be removed to
+    /// avoid creating a loop (where the Image is both being drawn *from* and being drawn *to*.)
+    pub fn detach(&mut self) -> Option<Image> {
+        Some(Image::new(self.0.take_texture()?))
+    }
+
+    /// Get the pixel data of a given region of this surface
+    pub fn screenshot(
+        &self,
+        gfx: &Graphics,
+        x: u32,
+        y: u32,
+        width: u32,
+        height: u32,
+        format: ColorFormat,
+    ) -> Vec<u8> {
+        self.0.bind();
+        let mut buffer = vec![0; (width * height * format.bytes_per_pixel()) as usize];
+        self.0
+            .get_pixel_data(x, y, width, height, format, &mut buffer[..]);
+        golem::Surface::unbind(&gfx.ctx);
+
+        buffer
     }
 }
 
@@ -514,288 +716,3 @@ fn join_change_lists<'a, U, V>(
         }
     })
 }
-
-const CIRCLE_LEN: usize = 63;
-
-// Until there's serious compile-time calculations in Rust,
-// it's best to just pre-write the points on a rasterized circle
-
-// Python script to generate the array:
-/*
-import math
-from math import pi
-
-
-def points_on_circumference(center=(0, 0), r=50, n=100):
-    n -= 1
-    return [
-        (
-            center[0]+(math.cos(2 * pi / n * x) * r),  # x
-            center[1] + (math.sin(2 * pi / n * x) * r)  # y
-
-        ) for x in range(0, n + 1)]
-
-
-number_of_points = 64
-points = points_on_circumference(center=(0,0),r=1, n=number_of_points)
-
-print("const CIRCLE_POINTS: [Vector; %i] = [" % number_of_points)
-
-for point in points:
-    print("    Vector { x: %s, y: %s }," % (point[0], point[1]))
-
-print("];")
-*/
-#[allow(clippy::unreadable_literal)]
-#[allow(clippy::excessive_precision)]
-const CIRCLE_POINTS: [Vector; CIRCLE_LEN] = [
-    Vector { x: 1.0, y: 0.0 },
-    Vector {
-        x: 0.9950307753654014,
-        y: 0.09956784659581666,
-    },
-    Vector {
-        x: 0.9801724878485438,
-        y: 0.19814614319939758,
-    },
-    Vector {
-        x: 0.9555728057861407,
-        y: 0.2947551744109042,
-    },
-    Vector {
-        x: 0.9214762118704076,
-        y: 0.38843479627469474,
-    },
-    Vector {
-        x: 0.8782215733702285,
-        y: 0.47825397862131824,
-    },
-    Vector {
-        x: 0.8262387743159949,
-        y: 0.5633200580636221,
-    },
-    Vector {
-        x: 0.766044443118978,
-        y: 0.6427876096865394,
-    },
-    Vector {
-        x: 0.6982368180860729,
-        y: 0.7158668492597184,
-    },
-    Vector {
-        x: 0.6234898018587336,
-        y: 0.7818314824680298,
-    },
-    Vector {
-        x: 0.5425462638657594,
-        y: 0.8400259231507714,
-    },
-    Vector {
-        x: 0.4562106573531629,
-        y: 0.8898718088114687,
-    },
-    Vector {
-        x: 0.365341024366395,
-        y: 0.9308737486442042,
-    },
-    Vector {
-        x: 0.27084046814300516,
-        y: 0.962624246950012,
-    },
-    Vector {
-        x: 0.17364817766693022,
-        y: 0.9848077530122081,
-    },
-    Vector {
-        x: 0.07473009358642417,
-        y: 0.9972037971811801,
-    },
-    Vector {
-        x: -0.024930691738072913,
-        y: 0.9996891820008162,
-    },
-    Vector {
-        x: -0.12434370464748516,
-        y: 0.9922392066001721,
-    },
-    Vector {
-        x: -0.22252093395631434,
-        y: 0.9749279121818236,
-    },
-    Vector {
-        x: -0.31848665025168454,
-        y: 0.9479273461671317,
-    },
-    Vector {
-        x: -0.41128710313061156,
-        y: 0.9115058523116731,
-    },
-    Vector {
-        x: -0.5000000000000002,
-        y: 0.8660254037844385,
-    },
-    Vector {
-        x: -0.58374367223479,
-        y: 0.8119380057158564,
-    },
-    Vector {
-        x: -0.6616858375968595,
-        y: 0.7497812029677341,
-    },
-    Vector {
-        x: -0.7330518718298263,
-        y: 0.6801727377709194,
-    },
-    Vector {
-        x: -0.7971325072229225,
-        y: 0.6038044103254774,
-    },
-    Vector {
-        x: -0.8532908816321556,
-        y: 0.5214352033794981,
-    },
-    Vector {
-        x: -0.900968867902419,
-        y: 0.43388373911755823,
-    },
-    Vector {
-        x: -0.9396926207859084,
-        y: 0.3420201433256685,
-    },
-    Vector {
-        x: -0.969077286229078,
-        y: 0.24675739769029342,
-    },
-    Vector {
-        x: -0.9888308262251285,
-        y: 0.14904226617617428,
-    },
-    Vector {
-        x: -0.9987569212189223,
-        y: 0.04984588566069704,
-    },
-    Vector {
-        x: -0.9987569212189223,
-        y: -0.04984588566069723,
-    },
-    Vector {
-        x: -0.9888308262251285,
-        y: -0.14904226617617447,
-    },
-    Vector {
-        x: -0.969077286229078,
-        y: -0.24675739769029362,
-    },
-    Vector {
-        x: -0.9396926207859084,
-        y: -0.34202014332566866,
-    },
-    Vector {
-        x: -0.9009688679024191,
-        y: -0.433883739117558,
-    },
-    Vector {
-        x: -0.8532908816321555,
-        y: -0.5214352033794983,
-    },
-    Vector {
-        x: -0.7971325072229224,
-        y: -0.6038044103254775,
-    },
-    Vector {
-        x: -0.7330518718298262,
-        y: -0.6801727377709195,
-    },
-    Vector {
-        x: -0.6616858375968594,
-        y: -0.7497812029677342,
-    },
-    Vector {
-        x: -0.5837436722347898,
-        y: -0.8119380057158565,
-    },
-    Vector {
-        x: -0.4999999999999996,
-        y: -0.8660254037844388,
-    },
-    Vector {
-        x: -0.4112871031306116,
-        y: -0.9115058523116731,
-    },
-    Vector {
-        x: -0.3184866502516841,
-        y: -0.9479273461671318,
-    },
-    Vector {
-        x: -0.2225209339563146,
-        y: -0.9749279121818236,
-    },
-    Vector {
-        x: -0.12434370464748495,
-        y: -0.9922392066001721,
-    },
-    Vector {
-        x: -0.024930691738073156,
-        y: -0.9996891820008162,
-    },
-    Vector {
-        x: 0.07473009358642436,
-        y: -0.9972037971811801,
-    },
-    Vector {
-        x: 0.17364817766693083,
-        y: -0.984807753012208,
-    },
-    Vector {
-        x: 0.2708404681430051,
-        y: -0.962624246950012,
-    },
-    Vector {
-        x: 0.3653410243663954,
-        y: -0.9308737486442041,
-    },
-    Vector {
-        x: 0.45621065735316285,
-        y: -0.8898718088114687,
-    },
-    Vector {
-        x: 0.5425462638657597,
-        y: -0.8400259231507713,
-    },
-    Vector {
-        x: 0.6234898018587334,
-        y: -0.7818314824680299,
-    },
-    Vector {
-        x: 0.698236818086073,
-        y: -0.7158668492597183,
-    },
-    Vector {
-        x: 0.7660444431189785,
-        y: -0.6427876096865389,
-    },
-    Vector {
-        x: 0.8262387743159949,
-        y: -0.563320058063622,
-    },
-    Vector {
-        x: 0.8782215733702288,
-        y: -0.4782539786213178,
-    },
-    Vector {
-        x: 0.9214762118704076,
-        y: -0.38843479627469474,
-    },
-    Vector {
-        x: 0.9555728057861408,
-        y: -0.2947551744109039,
-    },
-    Vector {
-        x: 0.9801724878485438,
-        y: -0.19814614319939772,
-    },
-    Vector {
-        x: 0.9950307753654014,
-        y: -0.09956784659581641,
-    },
-];